Move bit64/bit32 etc. from random_64bit_wrapper to random_64bit_generator
