+++ inst/include/dqrng_generator.h
@@ -39,20 +39,6 @@ class random_64bit_wrapper : public random_64bit_generator {
   static_assert(RNG::min() == 0, "Provided RNG has wrong minimum.");
 private:
   RNG gen;
-  bool has_cache{false};
-  uint32_t cache;
-
-  uint64_t bit64() {return gen();}
-  uint32_t bit32() {
-    if (has_cache) {
-      has_cache = false;
-      return cache;
-    }
-    uint64_t random = this->bit64();
-    cache = uint32_t(random);
-    has_cache = true;
-    return random >> 32;
-  }
   void set_stream(result_type stream) {throw std::runtime_error("Stream handling not supported for this RNG!");}
 
 protected:
@@ -64,7 +50,7 @@ public:
   random_64bit_wrapper(RNG _gen) : gen(_gen) {};
   random_64bit_wrapper(result_type seed) : gen(seed) {};
   random_64bit_wrapper(result_type seed, result_type stream) : gen(seed, stream) {};
-  virtual result_type operator() () override {return this->bit64();}
+  virtual result_type operator() () override {return gen();}
   virtual void seed(result_type seed) override {cache = false; gen.seed(seed);}
   virtual void seed(result_type seed, result_type stream) override {throw std::runtime_error("Stream handling not supported for this RNG!");}
   virtual std::unique_ptr<random_64bit_generator> clone(result_type stream) override {
@@ -72,78 +58,6 @@ public:
     rng->set_stream(stream);
     return rng;
   }
-
-  /*
-   * https://raw.githubusercontent.com/imneme/bounded-rands/3d71f53c975b1e5b29f2f3b05a74e26dab9c3d84/bounded32.cpp
-   * https://raw.githubusercontent.com/imneme/bounded-rands/3d71f53c975b1e5b29f2f3b05a74e26dab9c3d84/bounded64.cpp
-   * A C++ implementation methods and benchmarks for random numbers in a range
-   * (64 and 32-bit version)
-   *
-   * The MIT License (MIT)
-   *
-   * Copyright (c) 2018 Melissa E. O'Neill
-   *
-   * Permission is hereby granted, free of charge, to any person obtaining a
-   * copy of this software and associated documentation files (the "Software"),
-   * to deal in the Software without restriction, including without limitation
-   * the rights to use, copy, modify, merge, publish, distribute, sublicense,
-   * and/or sell copies of the Software, and to permit persons to whom the
-   * Software is furnished to do so, subject to the following conditions:
-   *
-   * The above copyright notice and this permission notice shall be included in
-   * all copies or substantial portions of the Software.
-   *
-   * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-   * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-   * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-   * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-   * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-   * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-   * DEALINGS IN THE SOFTWARE.
-   */
-
-  virtual uint32_t operator() (uint32_t range) override {
-    uint32_t x = this->bit32();
-    uint64_t m = uint64_t(x) * uint64_t(range);
-    uint32_t l = uint32_t(m);
-    if (l < range) {
-      uint32_t t = -range;
-      if (t >= range) {
-        t -= range;
-        if (t >= range)
-          t %= range;
-      }
-      while (l < t) {
-        x = this->bit32();
-        m = uint64_t(x) * uint64_t(range);
-        l = uint32_t(m);
-      }
-    }
-    return m >> 32;
-  }
-
-#ifdef LONG_VECTOR_SUPPORT
-  virtual uint64_t operator() (uint64_t range) override {
-    using pcg_extras::pcg128_t;
-    uint64_t x = this->bit64();
-    pcg128_t m = pcg128_t(x) * pcg128_t(range);
-    uint64_t l = uint64_t(m);
-    if (l < range) { // # nocov start
-      uint64_t t = -range;
-      if (t >= range) {
-        t -= range;
-        if (t >= range)
-          t %= range;
-      }
-      while (l < t) {
-        x = this->bit64();
-        m = pcg128_t(x) * pcg128_t(range);
-        l = uint64_t(m);
-      }
-    } // # nocov end
-    return m >> 64;
-  }
-#endif
 };
 
 template<>
