+++ inst/include/dqrng_types.h
@@ -24,13 +24,31 @@
 #include <stdexcept>
 #include <memory>
 #include <Rcpp/XPtr.h>
+#include <pcg_extras.hpp>
 
 namespace dqrng {
 
 class random_64bit_generator {
+private:
+  uint64_t bit64() {return this->operator()();}
+  uint32_t bit32() {
+    if (has_cache) {
+      has_cache = false;
+      return cache;
+    }
+    uint64_t random = this->bit64();
+    cache = uint32_t(random);
+    has_cache = true;
+    return random >> 32;
+  }
+
 protected:
+  bool has_cache{false};
+  uint32_t cache;
+
   virtual void output(std::ostream& ost) const = 0;
   virtual void input(std::istream& ist) = 0;
+
 public:
   using result_type = uint64_t;
 
@@ -41,9 +59,76 @@ public:
   virtual std::unique_ptr<random_64bit_generator> clone(result_type stream) = 0;
   static constexpr result_type min() {return 0;};
   static constexpr result_type max() {return UINT64_MAX;};
-  virtual uint32_t operator() (uint32_t range) = 0;
+  /*
+   * https://raw.githubusercontent.com/imneme/bounded-rands/3d71f53c975b1e5b29f2f3b05a74e26dab9c3d84/bounded32.cpp
+   * https://raw.githubusercontent.com/imneme/bounded-rands/3d71f53c975b1e5b29f2f3b05a74e26dab9c3d84/bounded64.cpp
+   * A C++ implementation methods and benchmarks for random numbers in a range
+   * (64 and 32-bit version)
+   *
+   * The MIT License (MIT)
+   *
+   * Copyright (c) 2018 Melissa E. O'Neill
+   *
+   * Permission is hereby granted, free of charge, to any person obtaining a
+   * copy of this software and associated documentation files (the "Software"),
+   * to deal in the Software without restriction, including without limitation
+   * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+   * and/or sell copies of the Software, and to permit persons to whom the
+   * Software is furnished to do so, subject to the following conditions:
+   *
+   * The above copyright notice and this permission notice shall be included in
+   * all copies or substantial portions of the Software.
+   *
+   * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+   * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+   * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+   * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+   * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+   * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+   * DEALINGS IN THE SOFTWARE.
+   */
+
+  uint32_t operator() (uint32_t range) {
+    uint32_t x = this->bit32();
+    uint64_t m = uint64_t(x) * uint64_t(range);
+    uint32_t l = uint32_t(m);
+    if (l < range) {
+      uint32_t t = -range;
+      if (t >= range) {
+        t -= range;
+        if (t >= range)
+          t %= range;
+      }
+      while (l < t) {
+        x = this->bit32();
+        m = uint64_t(x) * uint64_t(range);
+        l = uint32_t(m);
+      }
+    }
+    return m >> 32;
+  }
+
 #ifdef LONG_VECTOR_SUPPORT
-  virtual uint64_t operator() (uint64_t range) = 0;
+  uint64_t operator() (uint64_t range) {
+    using pcg_extras::pcg128_t;
+    uint64_t x = this->bit64();
+    pcg128_t m = pcg128_t(x) * pcg128_t(range);
+    uint64_t l = uint64_t(m);
+    if (l < range) { // # nocov start
+      uint64_t t = -range;
+      if (t >= range) {
+        t -= range;
+        if (t >= range)
+          t %= range;
+      }
+      while (l < t) {
+        x = this->bit64();
+        m = pcg128_t(x) * pcg128_t(range);
+        l = uint64_t(m);
+      }
+    } // # nocov end
+    return m >> 64;
+  }
 #endif
   friend std::ostream& operator<<(std::ostream& ost, const random_64bit_generator& e) {
     e.output(ost);
@@ -88,17 +173,7 @@ public:
   virtual std::unique_ptr<random_64bit_generator> clone(result_type stream) override {
     return gen->clone(stream);
   };
-
-  virtual uint32_t operator() (uint32_t range) override {
-    return (*gen)(range);
-  }
-#ifdef LONG_VECTOR_SUPPORT
-  virtual uint64_t operator() (uint64_t range) override {
-    return (*gen)(range);
-  }
-#endif
 };
-
 
 } // namespace dqrng
 #endif // DQRNG_TYPES_H
